fastapi_server_postgresql/requirements.txt
fastapi>=0.116.1
uvicorn[standard]>=0.35.0
pydantic>=2.11.0,<3.0.0
python-multipart>=0.0.18,<1.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
